fix: parse 5G-TMSI in Identity Response as unsigned 32-bit
TS 23.003 2.10.1 defines the 5G-TMSI as a full 32-bit value, but
HandleIdentityResponse parsed it with strconv.ParseInt(..., 16, 32), which
rejects values with the high bit set (>= 0x80000000) as out of range. Parse
unsigned and store the bit pattern in the int32 field (this AMF's own
allocator only assigns 1..MaxInt32, so GUTI construction is unaffected),
and format the OAM active-UE listing's TMSI via uint32 so a high-bit value
round-trips as 8 hex chars instead of Go's sign+magnitude form. Regression
test added.

+++ producer/oam.go
@@ -116,7 +116,7 @@ func HandleOAMActiveUEContextsFromDB(request *httpwrapper.Request) *httpwrapper.
 			Mcc:        ue.Tai.PlmnId.Mcc,
 			Mnc:        ue.Tai.PlmnId.Mnc,
 			Tac:        ue.Tai.Tac,
-			Tmsi:       fmt.Sprintf("%08x", ue.GetTmsi()),
+			Tmsi:       fmt.Sprintf("%08x", uint32(ue.GetTmsi())),
 		}
 		if ue.RanUe != nil && ue.RanUe[models.ACCESSTYPE__3_GPP_ACCESS] != nil {
 			ueContext.RanUeNgapId = ue.RanUe[models.ACCESSTYPE__3_GPP_ACCESS].RanUeNgapId
